% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{activate}
\alias{activate}
\title{activate}
\usage{
activate()
}
\value{
No return value, called for side effects
}
\description{
Activates the vendored package by adding the \code{./rvendor} directory to the search path.
This script can often be included in a .Rprofile script to activate automatically at shell activation.
}
\examples{
\dontrun{
rvendor::activate()
}
}
