% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{install}
\usage{
install(...)
}
\arguments{
\item{...}{Unused arguments, reserved for future expansion. If any arguments
are matched to \code{...}, renv will signal an error.}
}
\description{
Installs the vendored package to the \code{./rvendor} directory
}
\examples{
\dontrun{
rvendor::install("../<my-local-path>")
rvendor::install("<org-name>/<repo-name>")
}
}
