% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{uninstall}
\alias{uninstall}
\title{uninstall}
\usage{
uninstall(package_name)
}
\value{
No return value, called for side effects
}
\description{
Uninstalls the vendored package by removing it from the \code{./rvendor} directory.
}
\examples{
\dontrun{
rvendor::uninstall("<package-name>")
}
}
